Refactor configurable rule decorator to just be a function [#65]
