+++ src/assistants/midas.ts
@@ -1,10 +1,10 @@
+import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import inGame from "../engine/rules/inGame";
 import Item from "../gsi-data-classes/Item";
 import PlayerItems from "../gsi-data-classes/PlayerItems";
 import Rule from "../engine/Rule";
-import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -70,5 +70,5 @@ export default [
         ],
     }),
 ]
-    .map((rule) => new RuleDecoratorConfigurable(configTopic, rule))
+    .map((rule) => configurable(configTopic, rule))
     .map(inGame);
