+++ src/assistants/neutralItemDigReminder.ts
@@ -1,3 +1,4 @@
+import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
@@ -5,7 +6,6 @@ import inGame from "../engine/rules/inGame";
 import Item from "../gsi-data-classes/Item";
 import PlayerItems from "../gsi-data-classes/PlayerItems";
 import Rule from "../engine/Rule";
-import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -79,4 +79,4 @@ export default [
     }),
 ]
     .map(inGame)
-    .map((rule) => new RuleDecoratorConfigurable(configTopic, rule));
+    .map((rule) => configurable(configTopic, rule));
