+++ src/assistants/newNeutralTokensTimer.ts
@@ -1,9 +1,9 @@
+import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
 import Rule from "../engine/Rule";
 import RuleDecoratorAtMinute from "../engine/rules/RuleDecoratorAtMinute";
-import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -19,7 +19,7 @@ export default helper.neutral.tierTimeInfo.map(
     (time) =>
         new RuleDecoratorAtMinute(
             time,
-            new RuleDecoratorConfigurable(
+            configurable(
                 configTopic,
                 new Rule({
                     label: rules.assistant.newNeutralTokens,
