+++ src/assistants/philosophersStone.ts
@@ -1,11 +1,11 @@
 /* eslint-disable max-statements */
+import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
 import inGame from "../engine/rules/inGame";
 import PlayerItems from "../gsi-data-classes/PlayerItems";
 import Rule from "../engine/Rule";
-import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -86,5 +86,5 @@ export default [
             ),
     }),
 ]
-    .map((rule) => new RuleDecoratorConfigurable(configTopic, rule))
+    .map((rule) => configurable(configTopic, rule))
     .map(inGame);
